requirements.txt
django-debug-toolbar==1.9.1
requests==2.18.4
